# specs/002-course-centric-academy/contracts/api.yaml
openapi: 3.0.3
info:
  title: Neolingus Academy API
  description: Course-centric language certification platform API
  version: 1.0.0
  contact:
    name: Neolingus Academy
    email: api@neolingus.com
servers:
  - url: https://neolingus.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /ai/tutor/chat:
    post:
      summary: AI tutor interaction
      description: Get AI tutoring assistance with streaming responses
      tags: [AI Tutoring]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AITutorRequest"
      responses:
        "200":
          description: AI tutor response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AITutorResponse"
            text/event-stream:
              schema:
                type: string
                description: Streaming AI response
components:
  schemas:
...
    AITutorRequest:
      type: object
      required:
        - course_id
        - message
      properties:
        course_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
          nullable: true
        message:
          type: string
        context_type:
          type: string
          enum: [general, session_specific, weakness_focused]
          default: general
    AITutorResponse:
      type: object
      properties:
        message:
          type: string
        suggestions:
          type: array
          items:
            type: string
        resources:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              url:
                type: string
              type:
                type: string
        context_updated:
          type: boolean
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...